+++ module/blog/view/edit/edit.php
@@ -21,6 +21,7 @@
 			<?php echo template::submit('blogEditSubmit', [
 				'value' => 'Publier'
 			]); ?>
+
 		</div>
 	</div>
 	<div class="row">
@@ -73,35 +74,74 @@
 		'value' => $this->getData(['module', $this->getUrl(0), $this->getUrl(2), 'content'])
 	]); ?>
 	<div class="row">
-		<div class="col6">
+		<div class="col12">
 			<div class="block">
 				<h4>Options de publication</h4>
-				<?php echo template::select('blogEditUserId', $module::$users, [
-					'label' => 'Auteur',
-					'selected' => $this->getUser('id')
-				]); ?>
-				<?php echo template::date('blogEditPublishedOn', [
-					'help' => 'L\'article n\'est visible qu\'après la date de publication prévue.',
-					'label' => 'Date de publication',
-					'value' => $this->getData(['module', $this->getUrl(0), $this->getUrl(2), 'publishedOn'])
+				<div class="row">
+					<div class="col4">
+						<?php echo template::select('blogEditCommentMaxlength', $module::$commentLength,[
+							'help' => 'Choix du nombre maximum de caractères pour chaque commentaire de l\'article, mise en forme html comprise.',
+							'label' => 'Caractères par commentaire',
+							'selected' => $this->getData(['module', $this->getUrl(0), $this->getUrl(2), 'commentMaxlength'])
+						]); ?>
+					</div>
+					<div class="col4">
+						<?php echo template::select('blogEditUserId', $module::$users, [
+							'label' => 'Auteur',
+							'selected' => $this->getUser('id'),
+							'disabled' => $this->getUser('group') !== self::GROUP_ADMIN ? true : false
+						]); ?>
+					</div>
+					<div class="col4">
+						<?php echo template::date('blogEditPublishedOn', [
+							'help' => 'L\'article n\'est visible qu\'après la date de publication prévue.',
+							'label' => 'Date de publication',
+							'value' => $this->getData(['module', $this->getUrl(0), $this->getUrl(2), 'publishedOn'])
+						]); ?>
+					</div>
+				</div>
+			</div>
+		</div>
+	</div>
+	<div class="row">
+		<div class="col6">
+			<div class="block">
+			<h4>Permissions</h4>
+				<?php echo template::select('blogEditRights', $this->getUser('group') === self::GROUP_ADMIN ? $module::$articleRightsAdmin : $module::$articleRightsModerator , [
+					'label' => 'Droits d\'édition et de modification',
+					'selected' => $this->getData(['module', $this->getUrl(0), $this->getUrl(2), 'editRights'])
 				]); ?>
 			</div>
 		</div>
 		<div class="col6">
 			<div class="block">
-				<h4>Options avancées</h4>
+				<h4>Commentaires</h4>
 				<?php echo template::checkbox('blogEditCloseComment', true, 'Fermer les commentaires', [
 					'checked' => $this->getData(['module', $this->getUrl(0), $this->getUrl(2), 'closeComment'])
 				]); ?>
-				<?php echo template::checkbox('blogEditMailNotification', true, 'Notifier le commentaire aux groupes à partir de :', [
-					'checked' => $this->getData(['module', $this->getUrl(0), $this->getUrl(2), 'mailNotification']),
-					'help' => 'Editeurs = éditeurs + administrateurs<br/> Membres = membres + éditeurs + administrateurs'
-
-				]); ?>
-				<?php echo template::select('blogEditGroupNotification', $module::$groupNews, [
-					'label' => '',
-					'selected' => $this->getData(['module', $this->getUrl(0), $this->getUrl(2), 'groupNotification'])
-				]); ?>
+				<div id="commentOptionsWrapper">
+					<div class="row">
+						<div class="col12">
+							<?php echo template::checkbox('blogEditCommentApprove', true, 'Approbation des commentaires', [
+								'checked' => $this->getData(['module', $this->getUrl(0), $this->getUrl(2), 'commentApprove']),
+								''
+							]); ?>
+						</div>
+					</div>
+					<div class="row">
+						<div class="col7">
+							<?php echo template::checkbox('blogEditMailNotification', true, 'Notification des nouveaux commentaires par mail aux groupes', [
+								'checked' => $this->getData(['module', $this->getUrl(0), $this->getUrl(2), 'mailNotification']),
+							]); ?>
+						</div>
+						<div class="col5">
+							<?php echo template::select('blogEditGroupNotification', $module::$groupNews, [
+								'selected' => $this->getData(['module', $this->getUrl(0), $this->getUrl(2), 'groupNotification']),
+								'help' => 'Editeurs = éditeurs + administrateurs<br/> Membres = membres + éditeurs + administrateurs'
+							]); ?>
+						</div>
+					</div>
+				</div>
 			</div>
 		</div>
 	</div>