+++ module/blog/view/index/index.php
@@ -21,6 +21,7 @@
 						</a>
 					</div>
 					<div class="col9">
+						<article>
 						<h1 class="blogTitle">
 							<a href="<?php echo helper::baseUrl() . $this->getUrl(0) . '/' . $articleId; ?>">
 								<?php echo $article['title']; ?>
@@ -40,6 +41,7 @@
 							<?php echo helper::subword(strip_tags($article['content']), 0, 400); ?>...
 							<a href="<?php echo helper::baseUrl() . $this->getUrl(0) . '/' . $articleId; ?>">Lire la suite</a>
 						</p>
+						</article>
 					</div>
 				</div>
 			<?php endforeach; ?>